Modifications according to PR#32 comments
